process

+++ src/main/java/miu/edu/com/courseregistrationsystem/service/Registerer.java
@@ -1,5 +1,5 @@
 package miu.edu.com.courseregistrationsystem.service;
 
 public interface Registerer {
-    void process(long reventId);
+    void process(int reventId);
 }
